Added tests for FilmCard component, added utils folder
