+++ src/components/FilmCard/FilmCard.jsx
@@ -1,31 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const COLON = `:`;
-const COLON_REGEXP = /:/g;
-const SPACE = ` `;
-const LINE = `-`;
+import {getImgName} from '../../utils/utils';
 
 const FilmCard = (props) => {
-  const {film} = props;
-
-  const imgName = film
-    .split(SPACE)
-    .map((item) => {
-      if (item.includes(COLON)) {
-        const newName = item.replace(COLON_REGEXP, ``);
-        return newName.toLowerCase();
-      }
-      return item.toLowerCase();
-    })
-    .join(LINE);
+  const {film, onClick} = props;
 
   return (
     <article className="small-movie-card catalog__movies-card">
       <div className="small-movie-card__image">
-        <img src={`img/${imgName}.jpg`} alt={film} width="280" height="175"/>
+        <img src={`img/${getImgName(film)}.jpg`} alt={film} width="280" height="175"/>
       </div>
-      <h3 className="small-movie-card__title">
+      <h3 className="small-movie-card__title" onClick={onClick}>
         <a className="small-movie-card__link" href="movie-page.html">{film}</a>
       </h3>
     </article>
@@ -33,7 +19,8 @@ const FilmCard = (props) => {
 };
 
 FilmCard.propTypes = {
-  film: PropTypes.string.isRequired
+  film: PropTypes.string.isRequired,
+  onClick: PropTypes.func.isRequired
 };
 
 export default FilmCard;
