+++ src/components/app/App.jsx
@@ -5,7 +5,7 @@ import Main from '../main/Main';
 const films = [`Fantastic Beasts: The Crimes of Grindelwald`, `Bohemian Rhapsody`, `Macbeth`, `Aviator`, `We need to talk about Kevin`, `What We Do in the Shadows`, `Revenant`, `Johnny English`, `Shutter Island`, `Pulp Fiction`, `No Country for Old Men`, `Snatch`, `Moonrise Kingdom`, `Seven Years in Tibet`, `Midnight Special`, `War of the Worlds`, `Dardjeeling Limited`, `Orlando`, `Mindhunter`, `Midnight Special`];
 
 export const App = () => {
-  return <Main films={films} onClick={()=>{}}/>;
+  return <Main films={films} />;
 };
 
 export default App;
